requirements.txt
langchain==1.2.10
langchain-core==1.2.13
langchain-community==0.4.1
langchain-milvus==0.3.3
langchain-ollama==1.0.1
langchain-huggingface==1.2.0
langchain-text-splitters==0.3.0
langgraph==1.0.8
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.8.2
pydantic-settings==2.2.1
pymilvus==2.4.0
sentence-transformers==3.0.1
pyyaml==6.0.1
python-dotenv==1.0.0
rich==13.7.0
httpx==0.27.0

# Dev / testing
pytest==8.3.2
pytest-asyncio==0.24.0
